py-recognizer/requirements.txt
flask
pillow
numpy

